requirements.txt
discord.py
requests~=2.26.0
bs4~=0.0.1
discord-py-slash-command
discord-webhook
beautifulsoup4~=4.10.0
aiohttp~=3.7.4.post0
mysql-connector-python~=8.0.27